+++ flake.nix
@@ -1,9 +1,10 @@
 {
-  description = "CLI to manage passwords";
+  description = "CLI to manage passwords stored in OS-native keyrings, written in Rust";
 
   inputs = {
     nixpkgs = {
-      url = "github:nixos/nixpkgs/staging-next";
+      # until crates.io fix fully backported
+      url = "github:nixos/nixpkgs?tag=25.11&rev=c767db50e209f33ffce3c18165b36101079d367d";
     };
     fenix = {
       url = "github:nix-community/fenix/monthly";