+++ package.nix
@@ -17,7 +17,7 @@
 }:
 
 let
-  version = "0.1.0";
+  version = "1.0.0";
   hash = "";
   cargoHash = "";